CreateResource: convert data hash to HashWithIndifferentAccess
In Rails test environment, hashs are – in contrast to development env or
also the Acfs test environment– not automagically converted to strings. This
breakes the RTI check for the 'type'-attribute, since this is expected to be
a string.

For avoiding future errors while creating the resource, the data hash now gets
parsed as a ActiveSupport::HashWithIndifferentAccess, which can be accessed by
either stings or symbols as key.