Fix [RTI]: rescue NameError and NoMethodError on invalid type field value
When a `type` field is set in an Acfs-Resource, Acfs mimics Rails' STI by trying to instantiate a resource with the name of `type.camelize.constantize`. This fails, if the type string does not represent a valid subclass of the resource.
For better traceability, a `ResourceTypeError` indicating the type mismatch gets raised.