Add option to configure adapter creation
This commit adds an option to configure the adapter creation and to pass
keyword arguments to the typhoeus adapter. Arguments to the adapter will
be passed to Typhoeus::Hydra and can be sued to e.g. limit concurrency:

Acfs.configure do
  self.adapter = -> { # must be a proc
    Acfs::Adapter::Typhoeus.new max_concurrency: 20
  }
end

Please note that a new runner and adapter is created for each thread and
each thread uses its own typhoeus hydra. When limiting concurrency this
will only apply to one thread. E.g. limiting to 20 requests and running
four threads it will still be 80 parallel requests.