Fix Acfs.on callbacks for empty find_by results
Previously, Acfs.on callbacks would cause an error like this:

    undefined method `loaded?' for nil:Acfs::Util::ResourceDelegator

This happened because `find_by` tried to delegate to a `nil` object,
which can not stand in for a real Acfs resource (which has to
implement `loaded?`).

The commit adds checks for `nil` resources, which makes sure we do
not call `loaded?` on these objects anymore.